Tagging version 2.1.1

+++ release.sh
@@ -2,12 +2,17 @@
 
 VERSION=$1
 
-svn rm --force distr/latest_stable/*
-mkdir -p distr/latest_stable
-svn add distr/latest_stable
-cp -r trunk/netbeans-protobuf-plugin-suite/build/updates distr/$VERSION
-rm -rf distr/$VERSION/.svn
-cp trunk/netbeans-protobuf-plugin-suite/build/updates/* distr/latest_stable/
-svn add distr/$VERSION
-svn add distr/latest_stable/*
+if [ "$VERSION" == "" ]; then
+	exit
+else 
+	svn rm --force distr/latest_stable/*
+	mkdir -p distr/latest_stable
+	svn add distr/latest_stable
+	cp -r trunk/netbeans-protobuf-plugin-suite/build/updates distr/$VERSION
+	rm -rf distr/$VERSION/.svn
+	cp trunk/netbeans-protobuf-plugin-suite/build/updates/* distr/latest_stable/
+	svn add distr/$VERSION
+	svn add distr/latest_stable/*
 
+	echo "Remember to tag: $VERSION"
+fi
